Update: Modularise all SvgSymbol-related components
