+++ src/components/adslotUi/ListPickerComponent.jsx
@@ -8,7 +8,7 @@ import FlexibleSpacer from 'components/alexandria/FlexibleSpacer/component';
 import Grid from 'components/alexandria/Grid/component';
 import GridRow from 'components/alexandria/GridRow/component';
 import GridCell from 'components/alexandria/GridCell/component';
-import SvgSymbol from 'components/alexandria/SvgSymbolComponent';
+import SvgSymbol from 'components/alexandria/SvgSymbol/component';
 
 require('styles/adslotUi/ListPicker.scss');
