+++ src/components/adslotUi/PanelComponent.jsx
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react';
-import SvgSymbol from 'components/alexandria/SvgSymbolComponent';
+import SvgSymbol from 'components/alexandria/SvgSymbol/component';
 
 require('styles/adslotUi/Panel.scss');
