+++ src/components/adslotUi/TreePicker/TreePickerGridComponent.jsx
@@ -6,7 +6,7 @@ import TreePickerPropTypes from 'helpers/propTypes/TreePickerPropTypes';
 import Empty from 'components/alexandria/Empty/component';
 import Grid from 'components/alexandria/Grid/component';
 import GridRow from 'components/alexandria/GridRow/component';
-import SvgSymbol from 'components/alexandria/SvgSymbolComponent';
+import SvgSymbol from 'components/alexandria/SvgSymbol/component';
 
 require('styles/adslotUi/TreePickerGrid.scss');
