+++ src/components/alexandria/SvgSymbol/Circle/styles.scss
@@ -1,7 +1,7 @@
-@import '../border';
-@import '../color';
-@import '../function';
-@import '../variable';
+@import 'styles/border';
+@import 'styles/color';
+@import 'styles/function';
+@import 'styles/variable';
 
 .svgsymbolcircle-component {
   @mixin svg-size($diameter) {
