+++ src/components/alexandria/SvgSymbol/styles.scss
@@ -1,5 +1,5 @@
-@import '../color';
-@import '../variable';
+@import 'styles/color';
+@import 'styles/variable';
 
 .svg-symbol-component {
   @mixin size($diameter) {
