+++ src/components/core.js
@@ -53,8 +53,8 @@ import Search from 'components/alexandria/Search/component';
 import Slicey from 'components/alexandria/SliceyComponent';
 import Spinner from 'components/alexandria/Spinner/component';
 import Statistic from 'components/alexandria/Statistic/component';
-import SvgSymbol from 'components/alexandria/SvgSymbolComponent';
-import SvgSymbolCircle from 'components/alexandria/SvgSymbolCircleComponent';
+import SvgSymbol from 'components/alexandria/SvgSymbol/component';
+import SvgSymbolCircle from 'components/alexandria/SvgSymbol/Circle/component';
 import Tag from 'components/alexandria/Tag/component';
 import TileGrid from 'components/alexandria/TileGrid/component';
 import Totals from 'components/alexandria/Totals/component';
